test v1/requirements.txt
google-generativeai==0.3.1
streamlit==1.27.0
python-dotenv==1.0.0
tavily-python==0.2.6 